requirements.txt
fastapi==0.111.0
httpx==0.27.0
pydantic==2.7.4
pydantic_core==2.18.4
uvicorn==0.30.1
python-dotenv==1.0.1